+++ src/Security/SecurityInterface.php
@@ -27,9 +27,9 @@ interface SecurityInterface
 	/**
 	 * Get users Ip address.
 	 *
-	 * @param bool $secure Determine if the function will return normal IP or hashed IP.
+	 * @param string $secureType Determine if the function will return normal, hashed or anonymized IP.
 	 *
 	 * @return string
 	 */
-	public function getIpAddress(bool $secure = false): string;
+	public function getIpAddress(string $secureType = 'none'): string;
 }